requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
ultralytics==8.0.206
groq==0.11.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
streamlit==1.28.1
streamlit-chat==0.1.1
requests==2.31.0
aiofiles==23.2.0
